+++ config/autoload/redis.global.php
@@ -6,22 +6,19 @@ use Shlinkio\Shlink\Core\Config\EnvVars;
 
 return (static function (): array {
     $redisServers = EnvVars::REDIS_SERVERS->loadFromEnv();
-    $pubSub = [
+    $redis = ['pub_sub_enabled' => $redisServers !== null && EnvVars::REDIS_PUB_SUB_ENABLED->loadFromEnv(false)];
+    $cacheRedisBlock = $redisServers === null ? [] : [
         'redis' => [
-            'pub_sub_enabled' => $redisServers !== null && EnvVars::REDIS_PUB_SUB_ENABLED->loadFromEnv(false),
+            'servers' => $redisServers,
+            'sentinel_service' => EnvVars::REDIS_SENTINEL_SERVICE->loadFromEnv(),
         ],
     ];
 
-    return match ($redisServers) {
-        null => $pubSub,
-        default => [
-            'cache' => [
-                'redis' => [
-                    'servers' => $redisServers,
-                    'sentinel_service' => EnvVars::REDIS_SENTINEL_SERVICE->loadFromEnv(),
-                ],
-            ],
-            ...$pubSub,
+    return [
+        'cache' => [
+            'namespace' => 'Shlink',
+            ...$cacheRedisBlock,
         ],
-    };
+        'redis' => $redis,
+    ];
 })();